move theme to utils
