+++ packages/webcomponents/src/components/card-form/card-form.tsx
@@ -1,6 +1,6 @@
 import { Component, Event, Prop, h, EventEmitter, Method, Listen, State } from '@stencil/core';
 import { CreatePaymentMethodResponse } from '../payment-method-form/payment-method-responses';
-import { Theme } from '../payment-method-form/theme';
+import { Theme } from '../../utils/theme';
 
 @Component({
   tag: 'justifi-card-form',
@@ -48,7 +48,7 @@ export class CardForm {
    * Triggered when the validate method is called on the component
    * @event justifi-card-form#cardFormValidated
    */
-  @Event({ eventName: 'cardFormValidate'}) cardFormValidated: EventEmitter<{ data: { isValid: boolean } }>;
+  @Event({ eventName: 'cardFormValidate' }) cardFormValidated: EventEmitter<{ data: { isValid: boolean } }>;
 
   @Listen('paymentMethodFormReady')
   readyHandler(event: CustomEvent) {
